Add files via upload
This savegame as of 13.July 2019 has everything unlocked aside from alternate blueprints, for those it has 50 harddrives cheated in (there are only 39 alternate bp´s) 4 creatures have been killed and 1 carapace researched. Nothing has been build, only the HUB. My character is in the HUB and you just kill me. the 50 harddrives are in the box.